67 - Implement model ( https://italent.cloudapp.net:9443/ccm/web/projects/iTalent#action=com.ibm.team.workitem.viewWorkItem&id=67 )

+++ italent/src/main/java/be/italent/model/Milestone.java
@@ -16,7 +16,7 @@ public class Milestone extends AbstractITalentEntity implements Serializable {
 	private static final long serialVersionUID = 4572147731049218685L;
 	@Id
 	@GeneratedValue
-	private long id;
+	private int id;
 	
 	private boolean done;
 	
@@ -24,11 +24,11 @@ public class Milestone extends AbstractITalentEntity implements Serializable {
 	@ManyToOne
 	private Project project;
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
@@ -47,4 +47,6 @@ public class Milestone extends AbstractITalentEntity implements Serializable {
 	public void setProject(Project project) {
 		this.project = project;
 	}
+	
+	
 }
